% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allrules.R
\name{allrules}
\alias{allrules}
\title{Summary of the division rules}
\usage{
allrules(E, d, draw = TRUE)
}
\arguments{
\item{E}{The endowment.}

\item{d}{The vector of claims.}

\item{draw}{A logical value.}
}
\value{
A data-frame with the awards vectors selected by the main division rules. If draw = TRUE,
a bar chart for each claimant representing the amounts that the claimant receives.
}
\description{
This function returns the awards vectors selected, for a given claims problem, by the rules:  AA, APRO, CE, CEA, CEL, DT, MO, PIN, PRO, RA, and Talmud.
}
\details{
Let \eqn{E\ge 0} be the endowment to be divided and \eqn{d\in \mathcal{R}^n}{%
d} the vector of claims with \eqn{d\ge 0} and such that \eqn{\sum_{i=1}^{n} d_i\ge E,\ }{}
the sum of claims exceeds the endowment.

A vector \eqn{x=(x_1,\dots,x_n)}{x=(x1,...,xn)} is an awards vector for the claims problem \eqn{(E,d)} if:
no claimant is asked to pay (\eqn{0\le x});
no claimant  receives more than his claim (\eqn{x\le d});
and the balance requirement is satisfied, that is, the sum of the awards is equal to the endowment (\eqn{\sum_{i=1}^{n} x_i= E}{x1+...+xn=E}).

A rule is a function that assigns to each claims problem \eqn{(E,d)} an awards vector for \eqn{(E,d)},
that is, a division between the claimants of the amount available.

The formal definitions of the main rules are given in the corresponding function help.
}
\examples{
E=10
d=c(2,4,7,8)
allrules(E,d)
}
\references{
Mirás Calvo, M.A., Quinteiro Sandomingo, C., and Sánchez-Rodríguez, E. (2020). The core-center rule for the bankruptcy problem. Working paper 2020-02, ECOBAS.

Thomson, W. (2019). How to divide when there isn't enough. From Aristotle, the Talmud, and Maimonides to the axiomatics of resource allocation. Cambridge University Press.
}
\seealso{
\link{AA}, \link{APRO}, \link{CD}, \link{CE}, \link{CEA}, \link{CEL}
, \link{DT}, \link{MO}, \link{PIN}, \link{PRO}, \link{RA}, \link{Talmud},  \link{verticalruleplot}
}
