% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bankruptcy.R
\docType{data}
\name{Bankruptcy}
\alias{Bankruptcy}
\title{Bankruptcy data}
\format{
A data frame with 8 rows and 2 variables:


\tabular{rlll}{
[,1] \tab \code{Creditor}   \tab categorical \tab Creditor name \cr
[,2] \tab \code{Claim}  \tab numeric \tab Claim (millions of euro)\cr
}
}
\source{
The data were obtained from Fiestras et al (2016)
}
\usage{
Bankruptcy
}
\description{
Bankruptcy data: creditors and claims
}
\examples{
data(Bankruptcy)
Bankruptcy
E= 230
allrules(E,Bankruptcy$Claim)
}
\references{
Fiestras-Janeiro, M.G., Sánchez-Rodríguez, E., and Schuster, M. (2016). A precedence constraint value revisited. Top, 25, 156–179
}
\keyword{datasets}
