% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CE.R
\name{CE}
\alias{CE}
\title{Constrained egalitarian rule}
\usage{
CE(E, d, name = FALSE)
}
\arguments{
\item{E}{The endowment.}

\item{d}{The vector of claims.}

\item{name}{A logical value.}
}
\value{
The awards vector selected by the CE rule. If \code{name = TRUE}, the name of the function (CE) as a character string.
}
\description{
This function returns  the awards vector assigned by the constrained egalitarian rule (CE) rule to a claims problem.
}
\details{
Let \eqn{N=\{1,\ldots,n\}} be the set of claimants, \eqn{E\ge 0} the endowment to be divided and \eqn{d\in \mathbb{R}_+^N} the vector of claims
such that \eqn{D=\sum_{i \in N} d_i\ge E}.

Rearrange the claims from small to large, \eqn{0 \le d_1 \le...\le d_n}.
The constrained egalitarian rule (CE) coincides with the constrained equal awards rule (CEA) applied to
the problem \eqn{(E, d/2)} if the endowment is less or equal than the half-sum of the claims, \eqn{D/2}.
Otherwise, any additional unit is assigned to claimant \eqn{1} until she/he receives the minimum
of the claim and half of \eqn{d_2}. If this minimun is \eqn{d_1}, she/he stops there. If it is not, the
next increment is divided equally between claimants \eqn{1} and \eqn{2} until claimant \eqn{1} receives
 \eqn{d_1} (in this case she drops out) or they reach \eqn{d_3/2}.
If claimant \eqn{1} leaves, claimant \eqn{2} receives any additional increment until she/he reaches \eqn{d_2}
or \eqn{d_3/2}. In the case that claimant \eqn{1} and \eqn{2} reach \eqn{d_3/2}, any additional unit is
divided between claimants \eqn{1}, \eqn{2}, and \eqn{3} until the first one receives \eqn{d_1} or they
reach \eqn{d_4/2}, and so on. Therefore, for each \eqn{i\in N},

\deqn{\text{CE}_i(E,d)=\begin{cases}
\min\{\frac{d_i}{2},\lambda\}     & \text{if } E\leq \tfrac{1}{2}D\\[3pt]
\max \bigl\{ \frac{d_i}{2},\min\{d_i,\lambda \} \bigr\} & \text{if }  E \geq \tfrac{1}{2}D
\end{cases},}

where \eqn{\lambda \geq 0} is chosen such that  \eqn{\underset{i\in N}{\sum}	\text{CE}_i(E,d)=E}.
}
\examples{
E=10
d=c(2,4,7,8)
CE(E,d)
}
\references{
Chun, Y., Schummer, J., Thomson, W. (2001). Constrained egalitarianism: a new solution for claims problems. Seoul Journal of Economics 14, 269–297.

Thomson, W. (2019). How to divide when there isn't enough. From Aristotle, the Talmud, and Maimonides to the axiomatics of resource allocation. Cambridge University Press.
}
\seealso{
\link{allrules}, \link{axioms}, \link{CEA}, \link{PIN}, \link{Talmud}.
}
