% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Talmud.R
\name{Talmud}
\alias{Talmud}
\title{Talmud rule}
\usage{
Talmud(E, d, name = FALSE)
}
\arguments{
\item{E}{The endowment.}

\item{d}{The vector of claims.}

\item{name}{A logical value.}
}
\value{
The awards vector selected by the Talmud rule.
If \code{name = TRUE}, the name of the function (Talmud) as a character string.
}
\description{
This function returns the awards vector assigned by the Talmud rule to a claims problem.
}
\details{
Let \eqn{N=\{1,\ldots,n\}} be the set of claimants, \eqn{E\ge 0} the endowment to be divided and \eqn{d\in \mathbb{R}_+^N} the vector of claims
such that \eqn{D=\sum_{i \in N} d_i\ge E}.

The Talmud rule (Talmud) coincides with the constrained equal awards rule (CEA)
applied to the problem \eqn{(E, d/2)} if the endowment is less or equal than the half-sum of the claims, \eqn{D/2}.
Otherwise, the Talmud rule assigns \eqn{d/2} and
the remainder, \eqn{E-D/2}, is awarded with the constrained equal losses rule with claims \eqn{d/2}. Therefore, for each \eqn{i\in N},

\deqn{\text{Talmud}_i(E,d) = \begin{cases}
\min\{\frac{d_i}{2},\lambda\}   & \text{if }  E\leq \tfrac{1}{2}D\\[3pt]
d_i-\min\{\frac{d_i}{2},\lambda\} & \text{if }  E \geq \tfrac{1}{2}D
\end{cases},}

where \eqn{\lambda \geq 0} is chosen such that  \eqn{\underset{i\in N}{\sum}	\text{Talmud}_i(E,d)=E}.

The Talmud rule when applied to a two-claimant problem is often referred to as the contested garment rule and coincides with concede-and-divide rule.
The Talmud rule corresponds to the nucleolus of the associated (pessimistic) coalitional game.
}
\examples{
E=10
d=c(2,4,7,8)
Talmud(E,d)
D=sum(d)
#The Talmud rule is self-dual
d-Talmud(D-E,d)
}
\references{
Aumann, R. and Maschler, M. (1985). Game theoretic analysis of a bankruptcy problem from the Talmud. Journal of Economic Theory 36, 195-213.

Thomson, W. (2019). How to divide when there isn't enough. From Aristotle, the Talmud, and Maimonides to the axiomatics of resource allocation. Cambridge University Press.
}
\seealso{
\link{AA}, \link{allrules}, \link{APRO}, \link{axioms}, \link{CEA}, \link{CEL}, \link{CD}, \link{RA}, \link{RTalmud}.
}
