% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verticalruleplot.R
\name{verticalruleplot}
\alias{verticalruleplot}
\title{Vertical rule plot}
\usage{
verticalruleplot(E, d, Rules, col = NULL, legend = TRUE)
}
\arguments{
\item{E}{The endowment.}

\item{d}{The vector of claims}

\item{Rules}{The rules: AA, APRO, CE, CEA, CEL, DT, MO, PIN, PRO, RA, Talmud.}

\item{col}{The colors. If col=NULL then the sequence of default colors is:
c("red", "blue", "green", "yellow", "pink", "coral4", "darkgray", "burlywood3", "black", "darkorange", "darkviolet").}

\item{legend}{A logical value. The color legend is shown if legend=TRUE.}
}
\value{
This function represents the claims vector and the awards vector assigned by several rules as vertical segments.
}
\description{
For each claimant, it plots a vertical line with his claim and a point on the awards vector of the chosen rules.
}
\details{
Let \eqn{E\ge 0} be the endowment to be divided and \eqn{d\in \mathcal{R}^n}{%
d} the vector of claims with \eqn{d\ge 0} and such that \eqn{\sum_{i=1}^{n} d_i\ge E,\ }{}
the sum of claims exceeds the endowment.

A vector \eqn{x=(x_1,\dots,x_n)}{x=(x1,...,xn)} is an awards vector for the claims problem \eqn{(E,d)} if:
no claimant is asked to pay (\eqn{0\le x});
no claimant  receives more than his claim (\eqn{x\le d});
and the balance requirement is satisfied, that is, the sum of the awards is equal to the endowment (\eqn{\sum_{i=1}^{n} x_i= E}{x1+...+xn=E}).

A rule is a function that assigns to each claims problem \eqn{(E,d)} an awards vector for \eqn{(E,d)},
that is, a division between the claimants of the amount available.

The formal definitions of the main rules are given in the corresponding function help.
}
\examples{
E=10
d=c(2,4,7,8)
Rules=c(Talmud,RA,AA)
col=c("red","green","blue")
verticalruleplot(E,d,Rules,col)
}
\references{
Thomson, W. (2019). How to divide when there isn't enough. From Aristotle, the Talmud, and Maimonides to the axiomatics of resource allocation. Cambridge University Press.
}
\seealso{
\link{allrules}, \link{pathawards}, \link{pathawards3}, \link{schedrule}, \link{schedrules}
}
